Remove cluster chunk processor since nextChunk is fully implemented in the request classes
