+++ source/server/admin/BUILD
@@ -419,7 +419,6 @@ envoy_cc_library(
     srcs = ["clusters_request.cc"],
     hdrs = ["clusters_request.h"],
     deps = [
-        ":clusters_chunk_processor_lib",
         ":clusters_params_lib",
         "//envoy/buffer:buffer_interface",
         "//envoy/http:codes_interface",
@@ -429,6 +428,9 @@ envoy_cc_library(
         "//source/common/buffer:buffer_lib",
         "//source/common/common:logger_lib",
         "//source/common/http:headers_lib",
+        "//source/common/json:json_streamer_lib",
+        "//source/common/upstream:host_utility_lib",
+        "@envoy_api//envoy/admin/v3:pkg_cc_proto",
     ],
 )
 
@@ -444,24 +446,5 @@ envoy_cc_library(
         "//envoy/buffer:buffer_interface",
         "//envoy/http:codes_interface",
         "//envoy/http:query_params_interface",
-    ],
-)
-
-envoy_cc_library(
-    name = "clusters_chunk_processor_lib",
-    srcs = ["clusters_chunk_processor.cc"],
-    hdrs = ["clusters_chunk_processor.h"],
-    deps = [
-        "//envoy/buffer:buffer_interface",
-        "//envoy/common:pure_lib",
-        "//envoy/upstream:cluster_manager_interface",
-        "//envoy/upstream:upstream_interface",
-        "//source/common/common:logger_lib",
-        "//source/common/json:json_streamer_lib",
-        "//source/common/network:utility_lib",
-        "//source/common/upstream:host_utility_lib",
-        "@com_google_absl//absl/strings:string_view",
-        "@envoy_api//envoy/admin/v3:pkg_cc_proto",
-        "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
     ],
 )
